# [TraceQL] Remove unnecessary spanset layer, iterators, and predicates