--- a/tempodb/encoding/vparquet2/block_traceql.go
+++ b/tempodb/encoding/vparquet2/block_traceql.go
@@ -1757,6 +1757,9 @@ func (c *traceCollector) String() string {
 	return "traceCollector{}"
 }
 
+// KeepGroup is called once per trace and creates its final spanset
+// containing all the matching spans.  These spans already contain all span-level and
+// resource-level data.
 func (c *traceCollector) KeepGroup(res *parquetquery.IteratorResult) bool {
 	finalSpanset := getSpanset()
 	clear(c.traceAttrs)