**What this PR does**:
The changes are:
1. Remove the intermediate spansets in batchCollector. Now it updates the spans in-place on the way to the traceCollector, which still creates the final spanset.
2. Remove unnecessary OrPredicates and UnionIterators.
3. Various other cleanups and simplifications

Benchmarks
```
name                                              old time/op    new time/op    delta
BackendBlockTraceQL/spanAttNameNoMatch              3.94ms ± 2%    3.91ms ± 1%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch               35.2ms ± 1%    34.4ms ± 0%   -2.27%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValMatch                 39.4ms ± 1%    37.4ms ± 2%   -5.14%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         3.84ms ± 1%    3.77ms ± 7%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch           70.3ms ± 1%    63.3ms ± 1%  -10.01%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch    4.30ms ± 1%    4.26ms ± 2%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch       107ms ± 2%      98ms ± 0%   -8.11%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch          3.17ms ± 2%    3.15ms ± 2%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch           3.00ms ± 1%    3.02ms ± 2%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttValMatch              625ms ± 4%     591ms ± 3%   -5.39%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch     3.02ms ± 0%    3.02ms ± 1%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch       66.5ms ± 1%    66.4ms ± 4%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch                 699ms ± 0%     702ms ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/mixedValNoMatch                  673ms ± 2%     670ms ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd           3.15ms ± 1%    3.14ms ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr             574ms ± 1%     564ms ± 1%   -1.82%  (p=0.016 n=5+5)
BackendBlockTraceQL/mixedValBothMatch               3.01ms ± 1%    3.03ms ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/count                            2.76s ± 1%     2.48s ± 2%  -10.27%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             3.38s ± 1%     3.32s ± 2%   -1.83%  (p=0.032 n=5+5)

name                                              old speed      new speed      delta
BackendBlockTraceQL/spanAttNameNoMatch             814MB/s ± 2%   820MB/s ± 1%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch              219MB/s ± 1%   224MB/s ± 0%   +2.32%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValMatch               94.7MB/s ± 1%  99.8MB/s ± 2%   +5.43%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch        762MB/s ± 1%   776MB/s ± 7%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch          317MB/s ± 1%   352MB/s ± 1%  +11.13%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch   680MB/s ± 1%   687MB/s ± 2%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch     208MB/s ± 2%   227MB/s ± 0%   +8.81%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch         399MB/s ± 2%   401MB/s ± 2%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch          177MB/s ± 1%   177MB/s ± 2%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttValMatch           62.4MB/s ± 4%  66.0MB/s ± 3%   +5.68%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch    222MB/s ± 0%   223MB/s ± 1%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch      548MB/s ± 1%   550MB/s ± 3%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch              18.7MB/s ± 0%  18.6MB/s ± 1%     ~     (p=0.476 n=5+5)
BackendBlockTraceQL/mixedValNoMatch               19.4MB/s ± 2%  19.5MB/s ± 1%     ~     (p=0.452 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd          400MB/s ± 1%   402MB/s ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr          44.4MB/s ± 1%  45.3MB/s ± 1%   +1.86%  (p=0.016 n=5+5)
BackendBlockTraceQL/mixedValBothMatch              223MB/s ± 1%   222MB/s ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/count                         12.9MB/s ± 1%  14.4MB/s ± 2%  +11.43%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                          14.1MB/s ± 1%  14.4MB/s ± 2%   +1.89%  (p=0.032 n=5+5)

name                                              old MB_io/op   new MB_io/op   delta
BackendBlockTraceQL/spanAttNameNoMatch                3.21 ± 0%      3.21 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttValNoMatch                 7.71 ± 0%      7.71 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttValMatch                   3.73 ± 0%      3.73 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicNoMatch           2.93 ± 0%      2.93 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch             22.3 ± 0%      22.3 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch      2.93 ± 0%      2.93 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch        22.3 ± 0%      22.3 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttNameNoMatch            1.26 ± 0%      1.26 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch             0.53 ± 0%      0.53 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValMatch               39.0 ± 0%      39.0 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch       0.67 ± 0%      0.67 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicMatch         36.5 ± 0%      36.5 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedNameNoMatch                  13.1 ± 0%      13.1 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValNoMatch                   13.1 ± 0%      13.1 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchAnd             1.26 ± 0%      1.26 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchOr              25.5 ± 0%      25.5 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValBothMatch                 0.67 ± 0%      0.67 ± 0%     ~     (all equal)
BackendBlockTraceQL/count                             35.7 ± 0%      35.7 ± 0%     ~     (all equal)
BackendBlockTraceQL/desc                              47.7 ± 0%      47.7 ± 0%     ~     (all equal)

name                                              old alloc/op   new alloc/op   delta
BackendBlockTraceQL/spanAttNameNoMatch              1.37MB ± 0%    1.37MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch               92.2MB ± 1%    91.5MB ± 1%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttValMatch                 1.38MB ± 0%    1.38MB ± 0%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         1.59MB ± 0%    1.59MB ± 0%   +0.00%  (p=0.032 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch           3.83MB ± 0%    3.86MB ± 0%   +0.59%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch    1.80MB ± 0%    1.80MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch      7.94MB ± 0%    7.96MB ± 0%   +0.26%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch          1.27MB ± 0%    1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch           1.27MB ± 0%    1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValMatch              184MB ± 0%     182MB ± 1%   -0.77%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch     1.27MB ± 0%    1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch       2.19MB ± 0%    2.19MB ± 0%   -0.38%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch                 101MB ± 3%     100MB ± 3%     ~     (p=0.087 n=5+5)
BackendBlockTraceQL/mixedValNoMatch                  109MB ± 0%     109MB ± 1%     ~     (p=0.254 n=5+4)
BackendBlockTraceQL/mixedValMixedMatchAnd           1.27MB ± 0%    1.27MB ± 0%     ~     (p=0.095 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr            3.95MB ± 0%    3.96MB ± 0%     ~     (p=0.063 n=5+4)
BackendBlockTraceQL/mixedValBothMatch               1.27MB ± 0%    1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/count                            937MB ± 0%     898MB ± 1%   -4.16%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             690MB ± 0%     674MB ± 0%   -2.32%  (p=0.008 n=5+5)

name                                              old allocs/op  new allocs/op  delta
BackendBlockTraceQL/spanAttNameNoMatch               18.7k ± 0%     18.7k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch                19.0k ± 0%     19.0k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/spanAttValMatch                  19.1k ± 0%     19.1k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch          18.7k ± 0%     18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch            30.1k ± 0%     30.1k ± 0%   -0.30%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch     21.0k ± 0%     21.0k ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch       73.0k ± 0%     72.9k ± 0%   -0.13%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch           18.7k ± 0%     18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch            18.7k ± 0%     18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValMatch              1.69M ± 0%     1.68M ± 0%   -0.76%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch      18.7k ± 0%     18.7k ± 0%   +0.01%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch        24.0k ± 0%     24.0k ± 0%   -0.10%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch                 21.1k ± 0%     21.0k ± 0%   -0.07%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValNoMatch                  21.9k ± 0%     21.9k ± 0%   -0.09%  (p=0.016 n=5+4)
BackendBlockTraceQL/mixedValMixedMatchAnd            18.7k ± 0%     18.7k ± 0%   -0.01%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr             31.2k ± 0%     31.1k ± 0%   -0.35%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValBothMatch                18.7k ± 0%     18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/count                            10.1M ± 0%      9.9M ± 0%   -1.62%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             6.56M ± 0%     6.47M ± 0%   -1.39%  (p=0.008 n=5+5)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`