@mdisibio 2023-10-03
Switched back to type checks instead of strings in res.OtherEntries and benchmarks are more clearcut as expected.  I had this originally but went to strings because it seemed more "correct" but micro-benchmarks showed that type checks are about significantly faster.  Should probably do the same for the remaining key checks but don't want to increase the scope of this PR any further.

```
name                                              old time/op    new time/op     delta
BackendBlockTraceQL/spanAttNameNoMatch              3.44ms ± 2%     3.44ms ± 2%     ~     (p=0.841 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch               33.9ms ± 1%     33.6ms ± 0%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/spanAttValMatch                 38.3ms ± 1%     36.7ms ± 2%   -4.25%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         3.46ms ± 3%     3.46ms ± 1%     ~     (p=0.690 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch           68.0ms ± 1%     61.7ms ± 1%   -9.32%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch    3.85ms ± 2%     3.80ms ± 0%     ~     (p=0.190 n=5+4)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch       104ms ± 1%       97ms ± 1%   -6.38%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch          3.06ms ± 2%     3.07ms ± 3%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch           2.94ms ± 1%     2.95ms ± 2%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttValMatch              606ms ± 2%      588ms ± 1%   -3.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch     2.97ms ± 2%     2.94ms ± 1%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch       63.8ms ± 1%     62.8ms ± 2%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch                 689ms ± 1%      695ms ± 1%     ~     (p=0.095 n=5+5)
BackendBlockTraceQL/mixedValNoMatch                  665ms ± 1%      663ms ± 1%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd           3.11ms ± 2%     3.06ms ± 1%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr             556ms ± 1%      552ms ± 1%     ~     (p=0.095 n=5+5)
BackendBlockTraceQL/mixedValBothMatch               2.95ms ± 1%     2.97ms ± 3%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/count                            2.67s ± 4%      2.44s ± 3%   -8.84%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             3.30s ± 1%      3.23s ± 1%   -2.05%  (p=0.008 n=5+5)

name                                              old speed      new speed       delta
BackendBlockTraceQL/spanAttNameNoMatch             932MB/s ± 2%    933MB/s ± 2%     ~     (p=0.841 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch              228MB/s ± 1%    230MB/s ± 0%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/spanAttValMatch               97.4MB/s ± 1%  101.8MB/s ± 2%   +4.45%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch        847MB/s ± 3%    847MB/s ± 1%     ~     (p=0.690 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch          327MB/s ± 1%    361MB/s ± 1%  +10.28%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch   761MB/s ± 2%    769MB/s ± 0%     ~     (p=0.190 n=5+4)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch     214MB/s ± 1%    229MB/s ± 1%   +6.82%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch         412MB/s ± 2%    411MB/s ± 3%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch          181MB/s ± 1%    180MB/s ± 2%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttValMatch           64.4MB/s ± 2%   66.4MB/s ± 1%   +3.09%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch    226MB/s ± 2%    228MB/s ± 1%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch      572MB/s ± 1%    581MB/s ± 2%     ~     (p=0.056 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch              19.0MB/s ± 1%   18.8MB/s ± 1%     ~     (p=0.079 n=5+5)
BackendBlockTraceQL/mixedValNoMatch               19.6MB/s ± 1%   19.7MB/s ± 1%     ~     (p=0.460 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd          406MB/s ± 2%    412MB/s ± 1%     ~     (p=0.222 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr          45.9MB/s ± 1%   46.2MB/s ± 1%     ~     (p=0.087 n=5+5)
BackendBlockTraceQL/mixedValBothMatch              228MB/s ± 1%    226MB/s ± 3%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/count                         13.4MB/s ± 4%   14.7MB/s ± 3%   +9.68%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                          14.5MB/s ± 1%   14.8MB/s ± 1%   +2.09%  (p=0.008 n=5+5)

name                                              old MB_io/op   new MB_io/op    delta
BackendBlockTraceQL/spanAttNameNoMatch                3.21 ± 0%       3.21 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttValNoMatch                 7.71 ± 0%       7.71 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttValMatch                   3.73 ± 0%       3.73 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicNoMatch           2.93 ± 0%       2.93 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch             22.3 ± 0%       22.3 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch      2.93 ± 0%       2.93 ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch        22.3 ± 0%       22.3 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttNameNoMatch            1.26 ± 0%       1.26 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch             0.53 ± 0%       0.53 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValMatch               39.0 ± 0%       39.0 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch       0.67 ± 0%       0.67 ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttIntrinsicMatch         36.5 ± 0%       36.5 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedNameNoMatch                  13.1 ± 0%       13.1 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValNoMatch                   13.1 ± 0%       13.1 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchAnd             1.26 ± 0%       1.26 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValMixedMatchOr              25.5 ± 0%       25.5 ± 0%     ~     (all equal)
BackendBlockTraceQL/mixedValBothMatch                 0.67 ± 0%       0.67 ± 0%     ~     (all equal)
BackendBlockTraceQL/count                             35.7 ± 0%       35.7 ± 0%     ~     (all equal)
BackendBlockTraceQL/desc                              47.7 ± 0%       47.7 ± 0%     ~     (all equal)

name                                              old alloc/op   new alloc/op    delta
BackendBlockTraceQL/spanAttNameNoMatch              1.37MB ± 0%     1.37MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch               92.0MB ± 1%     91.7MB ± 1%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/spanAttValMatch                 1.38MB ± 0%     1.38MB ± 0%     ~     (p=0.310 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch         1.59MB ± 0%     1.59MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch           3.83MB ± 0%     3.86MB ± 0%   +0.60%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch    1.80MB ± 0%     1.80MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch      7.94MB ± 0%     7.96MB ± 0%   +0.25%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch          1.27MB ± 0%     1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch           1.27MB ± 0%     1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValMatch              184MB ± 0%      182MB ± 0%   -0.95%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch     1.27MB ± 0%     1.27MB ± 0%   +0.00%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch       2.20MB ± 0%     2.19MB ± 0%   -0.39%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch                 100MB ± 3%      103MB ± 0%     ~     (p=0.683 n=5+4)
BackendBlockTraceQL/mixedValNoMatch                  110MB ± 0%      109MB ± 1%     ~     (p=0.190 n=5+4)
BackendBlockTraceQL/mixedValMixedMatchAnd           1.27MB ± 0%     1.27MB ± 0%     ~     (p=0.151 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr            3.95MB ± 0%     3.96MB ± 0%     ~     (p=0.190 n=5+4)
BackendBlockTraceQL/mixedValBothMatch               1.27MB ± 0%     1.27MB ± 0%   +0.00%  (p=0.016 n=5+5)
BackendBlockTraceQL/count                            936MB ± 0%      898MB ± 1%   -4.08%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             690MB ± 0%      674MB ± 0%   -2.34%  (p=0.008 n=5+5)

name                                              old allocs/op  new allocs/op   delta
BackendBlockTraceQL/spanAttNameNoMatch               18.7k ± 0%      18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttValNoMatch                19.0k ± 0%      19.0k ± 0%     ~     (p=0.635 n=5+5)
BackendBlockTraceQL/spanAttValMatch                  19.1k ± 0%      19.1k ± 0%     ~     (p=0.333 n=4+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch          18.7k ± 0%      18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicMatch            30.1k ± 0%      30.1k ± 0%   -0.30%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch     21.0k ± 0%      21.0k ± 0%     ~     (all equal)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch       73.0k ± 0%      72.9k ± 0%   -0.14%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch           18.7k ± 0%      18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/resourceAttValNoMatch            18.7k ± 0%      18.7k ± 0%     ~     (p=1.000 n=5+5)
BackendBlockTraceQL/resourceAttValMatch              1.69M ± 0%      1.68M ± 0%   -0.77%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch      18.7k ± 0%      18.7k ± 0%   +0.01%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch        24.0k ± 0%      24.0k ± 0%   -0.11%  (p=0.016 n=5+4)
BackendBlockTraceQL/mixedNameNoMatch                 21.1k ± 0%      21.0k ± 0%   -0.05%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValNoMatch                  21.9k ± 0%      21.9k ± 0%   -0.10%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd            18.7k ± 0%      18.7k ± 0%   -0.01%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr             31.2k ± 0%      31.1k ± 0%   -0.35%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValBothMatch                18.7k ± 0%      18.7k ± 0%     ~     (all equal)
BackendBlockTraceQL/count                            10.1M ± 0%       9.9M ± 0%   -1.62%  (p=0.008 n=5+5)
BackendBlockTraceQL/desc                             6.56M ± 0%      6.47M ± 0%   -1.39%  (p=0.008 n=5+5)
```